+++ packages/components/src/disabled/index.tsx
@@ -1,23 +1,33 @@
 /**
  * External dependencies
  */
-import classnames from 'classnames';
+import type { HTMLProps } from 'react';
 
 /**
  * WordPress dependencies
  */
 import { useDisabled } from '@wordpress/compose';
-import { createContext } from '@wordpress/element';
+import { createContext, forwardRef } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
-import { StyledWrapper } from './styles/disabled-styles';
+import { disabledStyles } from './styles/disabled-styles';
 import type { DisabledProps } from './types';
 import type { WordPressComponentProps } from '../ui/context';
+import { useCx } from '../utils';
 
 const Context = createContext< boolean >( false );
 const { Consumer, Provider } = Context;
+
+// Extracting this ContentWrapper component in order to make it more explicit
+// the same 'ContentWrapper' component is needed so that React can reconcile
+// the dom correctly when switching between disabled/non-disabled (instead
+// of thrashing the previous DOM and therefore losing the form fields values).
+const ContentWrapper = forwardRef<
+	HTMLDivElement,
+	HTMLProps< HTMLDivElement >
+>( ( props, ref ) => <div { ...props } ref={ ref } /> );
 
 /**
  * `Disabled` is a component which disables descendant tabbable elements and prevents pointer interaction.
@@ -54,22 +64,30 @@ function Disabled( {
 	children,
 	isDisabled = true,
 	...props
-}: Omit< WordPressComponentProps< DisabledProps, 'div' >, 'ref' > ) {
+}: WordPressComponentProps< DisabledProps, 'div' > ) {
 	const ref = useDisabled();
-
+	const cx = useCx();
 	if ( ! isDisabled ) {
-		return <Provider value={ false }>{ children }</Provider>;
+		return (
+			<Provider value={ false }>
+				<ContentWrapper>{ children }</ContentWrapper>
+			</Provider>
+		);
 	}
 
 	return (
 		<Provider value={ true }>
-			<StyledWrapper
+			<ContentWrapper
 				ref={ ref }
-				className={ classnames( className, 'components-disabled' ) }
+				className={ cx(
+					disabledStyles,
+					className,
+					'components-disabled'
+				) }
 				{ ...props }
 			>
 				{ children }
-			</StyledWrapper>
+			</ContentWrapper>
 		</Provider>
 	);
 }