+++ packages/components/src/flex/flex/hook.ts
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { css, SerializedStyles } from '@emotion/react';
+import { css } from '@emotion/react';
 
 /**
  * WordPress dependencies
@@ -16,7 +16,7 @@ import { useContextSystem, WordPressComponentProps } from '../../ui/context';
 import { useResponsiveValue } from '../../ui/utils/use-responsive-value';
 import { space } from '../../ui/utils/space';
 import * as styles from '../styles';
-import { useCx, rtl } from '../../utils';
+import { useCx } from '../../utils';
 import type { FlexProps } from '../types';
 
 function useDeprecatedProps(
@@ -61,72 +61,24 @@ export function useFlex( props: WordPressComponentProps< FlexProps, 'div' > ) {
 		typeof direction === 'string' && direction.includes( 'reverse' );
 
 	const cx = useCx();
-	const rtlWatchResult = rtl.watch();
 
 	const classes = useMemo( () => {
-		const sx: {
-			Base?: SerializedStyles;
-			Items?: SerializedStyles;
-			WrapItems?: SerializedStyles;
-		} = {};
-
-		sx.Base = css( {
+		const base = css( {
 			alignItems: isColumn ? 'normal' : align,
 			flexDirection: direction,
 			flexWrap: wrap ? 'wrap' : undefined,
+			gap: space( gap ),
 			justifyContent: justify,
 			height: isColumn && expanded ? '100%' : undefined,
 			width: ! isColumn && expanded ? '100%' : undefined,
-			marginBottom: wrap ? `calc(${ space( gap ) } * -1)` : undefined,
 		} );
-
-		/**
-		 * Workaround to optimize DOM rendering.
-		 * We'll enhance alignment with naive parent flex assumptions.
-		 *
-		 * Trade-off:
-		 * Far less DOM less. However, UI rendering is not as reliable.
-		 */
-		sx.Items = css`
-			> * + *:not( marquee ) {
-				margin-top: ${ isColumn ? space( gap ) : undefined };
-				${ rtl( {
-					marginLeft:
-						! isColumn && ! isReverse ? space( gap ) : undefined,
-					marginRight:
-						! isColumn && isReverse ? space( gap ) : undefined,
-				} )() }
-			}
-		`;
-
-		sx.WrapItems = css`
-			> *:not( marquee ) {
-				margin-bottom: ${ space( gap ) };
-				${ rtl( {
-					marginLeft:
-						! isColumn && isReverse ? space( gap ) : undefined,
-					marginRight:
-						! isColumn && ! isReverse ? space( gap ) : undefined,
-				} )() }
-			}
-
-			> *:last-child:not( marquee ) {
-				${ rtl( {
-					marginLeft: ! isColumn && isReverse ? 0 : undefined,
-					marginRight: ! isColumn && ! isReverse ? 0 : undefined,
-				} )() }
-			}
-		`;
 
 		return cx(
 			styles.Flex,
-			sx.Base,
-			wrap ? sx.WrapItems : sx.Items,
+			base,
 			isColumn ? styles.ItemsColumn : styles.ItemsRow,
 			className
 		);
-		// rtlWatchResult is needed to refresh styles when the writing direction changes
-		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [
 		align,
 		className,
@@ -138,7 +90,6 @@ export function useFlex( props: WordPressComponentProps< FlexProps, 'div' > ) {
 		isReverse,
 		justify,
 		wrap,
-		rtlWatchResult,
 	] );
 
 	return { ...otherProps, className: classes, isColumn };