+++ packages/core-data/src/hooks/use-resource-permissions.ts
@@ -44,6 +44,8 @@ type ResourcePermissionsResolution< IdType > = [
 /**
  * Resolves resource permissions.
  *
+ * @since 6.1.0 Introduced in WordPress core.
+ *
  * @param  resource The resource in question, e.g. media.
  * @param  id       ID of a specific resource entry, if needed, e.g. 10.
  *