+++ test/native/__mocks__/styleMock.js
@@ -168,4 +168,7 @@ module.exports = {
 	'wp-block-list-item__list-item--default': {
 		fontSize: 16,
 	},
+	'wp-block-list-item__list-item-placeholder': {
+		color: '#87a6bc',
+	},
 };